Refactored duplicate api_services into a single file, removed duplicate logger.ts files
