+++ src/services/acspRegistrationService.ts
@@ -1,6 +1,6 @@
 import { Resource } from "@companieshouse/api-sdk-node";
 import logger from "../utils/logger";
-import { createPublicOAuthApiClient } from "./api/api_service";
+import { createPublicOAuthApiClient } from "./apiService";
 import { Session } from "@companieshouse/node-session-handler";
 import ApiClient from "@companieshouse/api-sdk-node/dist/client";
 import { ApiErrorResponse } from "@companieshouse/api-sdk-node/dist/services/resource";
