--- src/services/api-services.ts
@@ -1,20 +1,0 @@
-import { API_URL, CHS_API_KEY } from "../../src/utils/properties";
-import { createApiClient } from "@companieshouse/api-sdk-node";
-import ApiClient from "@companieshouse/api-sdk-node/dist/client";
-import { Session } from "@companieshouse/node-session-handler";
-import { createAndLogError } from "../utils/logger";
-import { AccessTokenKeys } from "@companieshouse/node-session-handler/lib/session/keys/AccessTokenKeys";
-import { SignInInfoKeys } from "@companieshouse/node-session-handler/lib/session/keys/SignInInfoKeys";
-import { SessionKey } from "@companieshouse/node-session-handler/lib/session/keys/SessionKey";
-
-export const createPublicApiKeyClient = (): ApiClient => {
-    return createApiClient(CHS_API_KEY, undefined, API_URL);
-};
-
-export const createPaymentApiClient = (session: Session, paymentUrl: string): ApiClient => {
-    const oAuth = session.data?.[SessionKey.SignInInfo]?.[SignInInfoKeys.AccessToken]?.[AccessTokenKeys.AccessToken];
-    if (oAuth) {
-        return createApiClient(undefined, oAuth, paymentUrl);
-    }
-    throw createAndLogError("Error getting session keys for creating public api client");
-};
