+++ src/services/apiService.ts
@@ -2,18 +2,17 @@ import { Session } from "@companieshouse/node-session-handler";
 import { SessionKey } from "@companieshouse/node-session-handler/lib/session/keys/SessionKey";
 import { SignInInfoKeys } from "@companieshouse/node-session-handler/lib/session/keys/SignInInfoKeys";
 import { AccessTokenKeys } from "@companieshouse/node-session-handler/lib/session/keys/AccessTokenKeys";
-import { API_URL, CHS_API_KEY } from "../../utils/properties";
-import { createAndLogError } from "../../common/__utils/logger";
+import { API_URL, CHS_API_KEY } from "../utils/properties";
+import { createAndLogError } from "../utils/logger";
 import { createApiClient } from "@companieshouse/api-sdk-node";
 import ApiClient from "@companieshouse/api-sdk-node/dist/client";
-import { THERE_IS_A_PROBLEM_ERROR } from "../../common/__utils/constants";
 
 export const createPublicOAuthApiClient = (session: Session): ApiClient => {
     const oAuth = session.data?.[SessionKey.SignInInfo]?.[SignInInfoKeys.AccessToken]?.[AccessTokenKeys.AccessToken];
     if (oAuth) {
         return createApiClient(undefined, oAuth, API_URL);
     }
-    throw createAndLogError(THERE_IS_A_PROBLEM_ERROR, "Error getting session keys for creating public api client");
+    throw createAndLogError("Error getting session keys for creating public api client");
 };
 
 export const createPublicApiKeyClient = (): ApiClient => {
@@ -22,4 +21,12 @@ export const createPublicApiKeyClient = (): ApiClient => {
 
 export const createPublicApiKeyClientWithToken = (token: string): ApiClient => {
     return createApiClient(undefined, token, API_URL);
+};
+
+export const createPaymentApiClient = (session: Session, paymentUrl: string): ApiClient => {
+    const oAuth = session.data?.[SessionKey.SignInInfo]?.[SignInInfoKeys.AccessToken]?.[AccessTokenKeys.AccessToken];
+    if (oAuth) {
+        return createApiClient(undefined, oAuth, paymentUrl);
+    }
+    throw createAndLogError("Error getting session keys for creating public api client");
 };
